fix(ci): python-coverage-shards — repair orphan/provisioner tests and line-limit overflows
Shared-ancestor orphan recovery and is_ancestor_of_head left RecordingGit/fake
queues stale; oversized first-party modules also broke the 1500-line maintainability gate.
