+++ tests/unit/node/test_provisioner_parts/test_provisioner_part_001.py
@@ -34,7 +34,6 @@ from awf.node.provisioner import (
     _egress_plan_decision,
     _egress_plan_destination_category,
     _positive_int,
-    _provision_base_commit,
     _provision_checkout_base_branch,
     _provision_local_branch_name,
     _provision_remote_push_branch,
@@ -1077,6 +1076,15 @@ class TestSuccess:
                 del workspace_id
                 return "h" * 40
 
+            async def is_ancestor_of_head(self, *, workspace_id: str, commit: str) -> bool:
+                del workspace_id, commit
+                # Preferred forge base is still an ancestor of the PR head tip.
+                return True
+
+            async def merge_base_with_head(self, *, workspace_id: str, commit: str) -> str | None:
+                del workspace_id, commit
+                return None
+
         git = _RecordingGit()
         provisioner = Provisioner(
             session_factory=session_factory,
@@ -1146,50 +1154,6 @@ class TestSuccess:
 
         assert _provision_checkout_base_branch(ws) == "refs/pull/278/head"
         assert _provision_remote_push_branch(ws) == "feature/fork-head"
-
-    @pytest.mark.unit
-    def test_existing_feature_pr_checkout_preserves_target_base_commit(self) -> None:
-        ws = Workspace(
-            repo_url="https://github.com/dimileeh/aira-web.git",
-            branch_base="development",
-            branch_name="awf/retry",
-            remote_push_branch="feature/existing-head",
-            task_kind="feature_branch_pr",
-            pr_url="https://github.com/dimileeh/aira-web/pull/278",
-            pr_number=278,
-            base_commit="b" * 40,
-        )
-
-        assert _provision_base_commit(ws, checked_out_head="h" * 40) == "b" * 40
-
-    @pytest.mark.unit
-    @pytest.mark.parametrize(
-        "preferred_is_ancestor, merge_base, expected",
-        [
-            (True, None, "c" * 40),
-            (True, "b" * 40, "c" * 40),
-            (False, "b" * 40, "b" * 40),
-            (False, "  " + "b" * 40 + "  ", "b" * 40),
-            (False, None, "c" * 40),
-            (False, "   ", "c" * 40),
-        ],
-    )
-    def test_retain_ancestor_base_commit_prefers_shared_history(
-        self,
-        preferred_is_ancestor: bool,
-        merge_base: str | None,
-        expected: str,
-    ) -> None:
-        from awf.node.provisioner import _retain_ancestor_base_commit
-
-        assert (
-            _retain_ancestor_base_commit(
-                "c" * 40,
-                preferred_is_ancestor=preferred_is_ancestor,
-                merge_base=merge_base,
-            )
-            == expected
-        )
 
     @pytest.mark.unit
     @pytest.mark.parametrize(
