+++ tests/unit/service/test_workspace_retry_port.py
@@ -1434,195 +1434,3 @@ async def test_retry_rejects_host_port_conflict_when_target_node_unknown(
                 provider_readiness_override_reason="no-node conflict scan test",
                 provider_environ={},
             )
-
-
-@pytest.mark.unit
-async def test_retry_runtime_gate_override_excludes_source_from_port_conflict(
-    factory: async_sessionmaker[AsyncSession],
-    tmp_path,
-) -> None:
-    """When ignore_source_runtime_check=True, the runtime-not-released gate
-    is skipped, and the source workspace is excluded from port conflict
-    scanning because the retry replaces the source. Even though the source
-    still holds host ports (no runtime release event), retry succeeds because
-    excluding the source avoids a false conflict with itself."""
-    settings = _settings_with_host_home(tmp_path)
-    req = _request_with_preflight_override()
-    companion_req = {
-        "name": "sidecar",
-        "repo_url": "git@github.com:example/sidecar.git",
-        "base_branch": "main",
-        "ports": [[5432, 5434]],
-    }
-    payload = req.model_dump(mode="python")
-    payload["companions"] = [companion_req]
-    req_with_companion = WorkspaceCreateRequest.model_validate(payload)
-
-    async with factory() as session:
-        source = await create_workspace_row(
-            session,
-            req_with_companion,
-            settings=settings,
-            provider_environ={},
-        )
-        await session.commit()
-
-    await _mark_failed(factory, source.id, release_runtime=False)
-
-    async with factory() as session:
-        retry = await retry_workspace_row(
-            session,
-            source.id,
-            settings=settings,
-            provider_readiness_override=True,
-            provider_readiness_override_reason="auto-retry port conflict",
-            provider_environ={},
-            ignore_source_runtime_check=True,
-        )
-        assert retry.new_workspace.id != source.id
-
-
-@pytest.mark.unit
-async def test_retry_runtime_gate_override_succeeds_when_source_runtime_released(
-    factory: async_sessionmaker[AsyncSession],
-    tmp_path,
-) -> None:
-    """When ignore_source_runtime_check=True and the source runtime has been
-    released, the retry can proceed because the source's host ports are no
-    longer held and there is no conflict."""
-    settings = _settings_with_host_home(tmp_path)
-    req = _request_with_preflight_override()
-    companion_req = {
-        "name": "sidecar",
-        "repo_url": "git@github.com:example/sidecar.git",
-        "base_branch": "main",
-        "ports": [[5432, 5434]],
-    }
-    payload = req.model_dump(mode="python")
-    payload["companions"] = [companion_req]
-    req_with_companion = WorkspaceCreateRequest.model_validate(payload)
-
-    async with factory() as session:
-        source = await create_workspace_row(
-            session,
-            req_with_companion,
-            settings=settings,
-            provider_environ={},
-        )
-        await session.commit()
-
-    await _mark_failed(factory, source.id, release_runtime=True)
-
-    async with factory() as session:
-        retry = await retry_workspace_row(
-            session,
-            source.id,
-            settings=settings,
-            provider_readiness_override=True,
-            provider_readiness_override_reason="auto-retry source released",
-            provider_environ={},
-            ignore_source_runtime_check=True,
-        )
-        assert retry.new_workspace.id != source.id
-
-
-@pytest.mark.unit
-async def test_retry_runtime_gate_override_succeeds_no_host_ports_runtime_not_released(
-    factory: async_sessionmaker[AsyncSession],
-    tmp_path,
-) -> None:
-    """When ignore_source_runtime_check=True and there are no host-port
-    claims, retry must succeed even when the source compose stack has not
-    been released because there are no ports to collide on."""
-    settings = _settings_with_host_home(tmp_path)
-    req = _request_with_preflight_override()
-
-    async with factory() as session:
-        source = await create_workspace_row(
-            session,
-            req,
-            settings=settings,
-            provider_environ={},
-        )
-        await session.commit()
-
-    await _mark_failed(factory, source.id, release_runtime=False)
-
-    async with factory() as session:
-        retry = await retry_workspace_row(
-            session,
-            source.id,
-            settings=settings,
-            provider_readiness_override=True,
-            provider_readiness_override_reason="auto-retry no ports",
-            provider_environ={},
-            ignore_source_runtime_check=True,
-        )
-        assert retry.new_workspace.id != source.id
-
-
-@pytest.mark.unit
-async def test_retry_defaults_unset_worker_node_to_local_for_legacy_source_hostname(
-    factory: async_sessionmaker[AsyncSession],
-    tmp_path,
-) -> None:
-    """An upgraded local install may have failed source rows stamped with the
-    old container hostname while current local workers default to "local".
-    When the source runtime is already released, the retry reservation must be
-    placed on "local" so the local scheduler can list and claim it."""
-    settings = Settings(
-        _env_file=None,
-        host_home=str(tmp_path / "home"),
-        docker_host="",
-    )
-    req = _request_with_preflight_override()
-    companion_req = {
-        "name": "sidecar",
-        "repo_url": "git@github.com:example/sidecar.git",
-        "base_branch": "main",
-        "ports": [[5432, 5434]],
-    }
-    payload = req.model_dump(mode="python")
-    payload["companions"] = [companion_req]
-    req_with_companion = WorkspaceCreateRequest.model_validate(payload)
-
-    async with factory() as session:
-        source = await create_workspace_row(
-            session,
-            req_with_companion,
-            settings=settings,
-            provider_environ={},
-        )
-        await session.commit()
-
-    await _mark_failed(factory, source.id, release_runtime=True)
-
-    async with factory() as session:
-        repo = WorkspaceRepository(session)
-        ws = await repo.get(source.id)
-        assert ws is not None
-        ws.node_id = "legacy-container-hostname"
-        reservations = await ResourceReservationRepository(session).list_for_workspace(
-            source.id,
-        )
-        assert len(reservations) == 1
-        assert reservations[0].node_id == "local"
-        await session.commit()
-
-    async with factory() as session:
-        retry = await retry_workspace_row(
-            session,
-            source.id,
-            settings=settings,
-            provider_readiness_override=True,
-            provider_readiness_override_reason="legacy local hostname normalization test",
-            provider_environ={},
-        )
-        await session.commit()
-
-    async with factory() as session:
-        retried_reservations = await ResourceReservationRepository(
-            session,
-        ).list_for_workspace(retry.new_workspace.id)
-        assert len(retried_reservations) == 1
-        assert retried_reservations[0].node_id == "local"
